added a poll button and testing the arduino poll command
Tweaking the arduino interface for the ADC, need to add the bitshift

RCJ, The ADC reading kinda works, but the format needs to be verified.

RCJ, Added a dtop level comment to all of the header files.

updated semantics from MA to MilliAmps, modified gui display so that grid current dsplay is mA.

RCJ, began to make modifications in order to support smaller files and more efficient data use.

RCJ, Removed the files common_utils.h/.cpp and tested the code.

RCJ, updated the readme to include the feature list